refactor: reorg some stuff into resources
Note: should probably move resources to common.
