+++ game/src/main_menu/plugin.rs
@@ -1,6 +1,6 @@
 use bevy::prelude::*;
 
-use crate::{despawn_with_tag, AppState};
+use crate::{despawn_with_tag, resources::AppState};
 
 use super::{
     components, screens,
