+++ game/src/resources/mod.rs
@@ -3,12 +3,15 @@
 //!
 //! All resources can be added with default values to the game using the `ElementalistResourcesPlugin`.
 
+mod app_state;
+mod cleanup;
 mod cursor_position;
 mod fonts;
 mod plugin;
 mod return_to_state;
 mod spritesheet;
 
+pub use app_state::AppState;
 pub use cursor_position::*;
 pub use plugin::ElementalistResourcesPlugin;
 pub use return_to_state::ReturnToState;
