+++ game/src/resources/plugin.rs
@@ -6,30 +6,61 @@ use game_library::{
     CursorPosition, SpellChoices,
 };
 
+use crate::app_systems;
+
 use super::{
-    cursor_position::update_cursor_position, fonts::set_initial_fonts,
-    spritesheet::load_spell_atlas, ReturnToState,
+    cleanup::cleanup_then_exit, cursor_position::update_cursor_position, fonts::set_initial_fonts,
+    spritesheet::load_spell_atlas, AppState, ReturnToState,
 };
 
+/// Elementalist resources plugin. This loads the resources needed which may not be
+/// specific to a single module. This includes:
+///
+/// - App state
+/// - System to display version and build information (top right)
+/// - System to load data from `game_data` directory
+/// - Fonts
+///    - System to set the initial fonts
+///    - System to change the fonts
+/// - Cursor position
+///    - System to update the cursor position
+/// - The settings database on disk
+/// - The spell atlas (the spritesheet for the spells)
+/// - The spell choices resource (this probably should be moved to the Player plugin)
+/// - A state to return to when the settings menu is closed
 pub struct ElementalistResourcesPlugin;
 
 impl Plugin for ElementalistResourcesPlugin {
     fn build(&self, app: &mut App) {
+        // ### ADD STATES HERE ###
+        app.add_state::<AppState>();
+
         // ### ADD EVENTS HERE ###
         app.add_event::<ChangeFont>();
 
         // ### ADD RESOURCES HERE ###
-        app.insert_resource(CursorPosition::default());
-        app.insert_resource(SpellChoices::default());
-        app.insert_resource(FontResource::default());
-        app.init_resource::<ReturnToState>();
-        app.insert_resource(PkvStore::new("games.nwest.one", "Elementalist"));
+        app.init_resource::<ReturnToState>()
+            .insert_resource(CursorPosition::default())
+            .insert_resource(SpellChoices::default())
+            .insert_resource(FontResource::default())
+            .insert_resource(PkvStore::new("games.nwest.one", "Elementalist"));
 
         // ### GRAPHICS RESOURCES ###
         app.add_systems(Startup, load_spell_atlas);
 
         // ### ADD SYSTEMS HERE ###
-        app.add_systems(PostStartup, set_initial_fonts);
-        app.add_systems(Update, (update_cursor_position, change_font));
+        app.add_systems(
+            Startup,
+            (
+                app_systems::add_game_descriptor,
+                app_systems::load_data_file_dir,
+            ),
+        )
+        .add_systems(Update, (update_cursor_position, change_font))
+        // This system is [`PostStartup`] because we need to load any font settings from
+        // disk before assigning the initial fonts handles.
+        .add_systems(PostStartup, set_initial_fonts)
+        // handle the cleanup state
+        .add_systems(OnEnter(AppState::CleanUp), cleanup_then_exit);
     }
 }
